adding and editing controls

+++ LibraryApp/src/YourLibraryApp/YourLibraryApp.Application/BookService.cs
@@ -1,40 +1,57 @@
-﻿using YourLibraryApp.Domain;
+﻿using System;
+using System.Collections.Generic;
+using YourLibraryApp.Domain;
 
-namespace YourLibraryApp.Application;
-
-
-// BookService.cs
-public class BookService : IBookService
+namespace YourLibraryApp.Application
 {
-    private readonly IBookRepository _bookRepository;
-
-    public BookService(IBookRepository bookRepository)
+    public class BookService : IBookService
     {
-        _bookRepository = bookRepository;
-    }
+        private readonly IBookRepository _bookRepository;
 
-    public IEnumerable<Book> GetAllBooks()
-    {
-        return _bookRepository.GetAllBooks();
-    }
+        public BookService(IBookRepository bookRepository)
+        {
+            _bookRepository = bookRepository;
+        }
 
-    public Book GetBookById(int id)
-    {
-        return _bookRepository.GetBookById(id);
-    }
+        public IEnumerable<Book> GetAllBooks()
+        {
+            return _bookRepository.GetAllBooks();
+        }
 
-    public void AddBook(Book book)
-    {
-        _bookRepository.AddBook(book);
-    }
+        public Book GetBookById(int id)
+        {
+            return _bookRepository.GetBookById(id);
+        }
 
-    public void UpdateBook(Book book)
-    {
-        _bookRepository.UpdateBook(book);
-    }
+        public void AddBook(Book book)
+        {
+            // İlişkili varlık kontrolü
+            if (book.Author != null && book.Author.Id <= 0)
+                throw new ArgumentException("Invalid author");
 
-    public void DeleteBook(int id)
-    {
-        _bookRepository.DeleteBook(id);
+            _bookRepository.AddBook(book);
+        }
+
+        public void UpdateBook(Book book)
+        {
+            // Kitap geçerliliği kontrolü
+            if (book == null)
+                throw new ArgumentNullException(nameof(book), "Book cannot be null");
+
+            // İlişkili varlık kontrolü
+            if (book.Author != null && book.Author.Id <= 0)
+                throw new ArgumentException("Invalid author");
+
+            _bookRepository.UpdateBook(book);
+        }
+
+        public void DeleteBook(int id)
+        {
+            // Geçerli kitap ID kontrolü
+            if (id <= 0)
+                throw new ArgumentException("Invalid book ID");
+
+            _bookRepository.DeleteBook(id);
+        }
     }
 }
